refactor: altera componente para utilizar CSS modules
