+++ src/app/components/ProgramasEspacos/CardBase/CardBase.module.css
@@ -1,0 +1,89 @@
+.cardContainer {
+  width: 50%;
+  height: 300px;
+  display: grid;
+  grid-template-columns: 50% 50%;
+  background-color: #ffffff;
+  margin: 40px auto;
+  border-radius: 30px;
+  align-items: center;
+}
+
+.cardContent {
+  padding-left: 20px;
+  display: flex;
+  width: 80%;
+  flex-direction: column;
+  justify-content: space-around;
+  align-items: start;
+  text-align: left;
+  height: 80%;
+  margin-right: 10px;
+}
+
+.cardTitle {
+  font-size: 45px;
+  font-family: "PP Formula Condensed";
+  font-weight: 400;
+  color: #1eb4e9;
+}
+
+.cardBody {
+  font-family: Poppins;
+  margin-bottom: 10px;
+}
+
+.cardButton {
+  width: 100px;
+  height: 40px;
+  background-color: #1eb4e9;
+  color: #ffffff;
+  font-family: Poppins;
+  border-radius: 13px;
+}
+
+.cardImage {
+  height: 100%;
+  width: 100%;
+  height: 300px;
+  object-fit: cover;
+  border-radius: 30px;
+}
+
+@media (max-width: 1028px) {
+  .cardContainer {
+    width: 90%;
+    height: 300px;
+  }
+
+  .cardImage {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    border-radius: 30px;
+  }
+
+  .cardContent {
+    width: 100%;
+    height: 100%;
+  }
+}
+
+@media (max-width: 1380px) {
+  .cardContainer {
+    width: 60%;
+    height: 300px;
+  }
+
+  .cardContent {
+    width: 100%;
+    height: 100%;
+  }
+
+  .cardImage {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    border-radius: 30px;
+  }
+}
